tests/wpt/web-platform-tests/html/dom/render-blocking/parser-blocking-script.tentative.html

<!DOCTYPE html>
<title>Parser-blocking script elements are implicitly render-blocking</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/test-render-blocking.js"></script>

<script>
// Add some renderable content before parser inserts body
document.documentElement.appendChild(document.createTextNode('text'));

// Test must be setup before the parser-blocking script
test_render_blocking(
    () => assert_equals(window.dummy, 1),
    'Parser-blocking script is evaluated');
</script>

<script src="support/dummy-1.js?pipe=trickle(d1)"></script>

<div>Some more text</div>
