LayoutTests/imported/w3c/web-platform-tests/html/dom/render-blocking/parser-blocking-script.tentative-expected.txt

text
Some more text

PASS Rendering is blocked before render-blocking resources are loaded
PASS Parser-blocking script is evaluated
